Replaces all instances of 'uint' that aren't associated with an array index with 'u32'.
The recommendation from the Rust community seems to be to use 'int' and
'uint' when referring to array sizes and indices and to use one of the
sized integers (e.g. 'i32', 'i64', 'u32', 'u64') everywhere else.  For
performance considerations on both 32-bit and 64-bit machines, 'i32' and
'u32' seem to be recommended over 'i64' and 'u64'.