Adds trait `AsVector`.
The `AsVector` trait provides a single method that converts `&self` to
`&Vector3D`.  This allows methods such as `dot` to work with any
combination of `Vector3D` and `Direction3D`, without requiring to
specifically call `as_vector()` on the `Direction3D` objects.
Unfortunately, it's not yet possible to implement the math traits
(`Add`, `Mul`, etc.) using the `AsVector` trait.